refactor: 구매금액 입력 범위 조정

+++ src/main/java/lotto/domain/Earning.java
@@ -4,11 +4,11 @@ public class Earning {
 	private final int PERCENTAGE = 100;
 	private String earningRate;
 
-	public Earning(int money, long totalCashPrize) {
+	public Earning(long money, long totalCashPrize) {
 		this.earningRate = calculateEarningRate(money, totalCashPrize);
 	}
 
-	private String calculateEarningRate(int money, long totalCashPrize) {
+	private String calculateEarningRate(long money, long totalCashPrize) {
 		double earningRate = ((double) totalCashPrize / money) * PERCENTAGE;
 		return String.format("%.1f", earningRate);
 	}
