+++ src/test/java/lotto/domain/EarningTest.java
@@ -9,7 +9,7 @@ class EarningTest {
 	@Test
 	@DisplayName("투입금액보다 상금이 적을때")
 	void calculateEarningRateTest() {
-		int money = 8000;
+		long money = 8000;
 		long totalCashPrize = 5000;
 		Earning earning = new Earning(money, totalCashPrize);
 		String result = "62.5";
@@ -20,7 +20,7 @@ class EarningTest {
 	@Test
 	@DisplayName("투입금액보다 상금이 많을때")
 	void calculateEarningRateTest2() {
-		int money = 6000;
+		long money = 6000;
 		long totalCashPrize = 55000;
 		Earning earning = new Earning(money, totalCashPrize);
 		String result = "916.7";
@@ -31,7 +31,7 @@ class EarningTest {
 	@Test
 	@DisplayName("상금이 없을때")
 	void calculateEarningRateTest3() {
-		int money = 6000;
+		long money = 6000;
 		long totalCashPrize = 0;
 		Earning earning = new Earning(money, totalCashPrize);
 		String result = "0.0";
